backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
click==8.3.1
exceptiongroup==1.3.1
fastapi==0.128.8
greenlet==3.3.1
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
SQLAlchemy==2.0.46
starlette==0.52.1
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0